Update the vagrant provisioning script

+++ script/vagrant_bootstrap.sh
@@ -19,33 +19,41 @@ apt-get -y install ntp git make python-dev python-pip libjpeg-dev libwebp-dev li
 pip install supervisor virtualenv
 
 # setup the supervisor configs
-echo ""
-echo "+------------------+"
-echo "| Setup Supervisor |"
-echo "+------------------+"
-echo ""
-echo_supervisord_conf > /etc/supervisord.conf
-cat /vagrant/script/supervisord > /etc/init.d/supervisord
-# add custom tweaks to config file
-sed -i 's/;chmod=0700/chmod=0764/' /etc/supervisord.conf
-sed -i 's/;chown=nobody:nogroup/chown=root:vagrant/' /etc/supervisord.conf
-sed -i 's/;\[include\]/\[include\]/' /etc/supervisord.conf
-echo "files = /home/vagrant/supervisor/*.conf" >> /etc/supervisord.conf
-su - vagrant -c 'mkdir ~/supervisor'
-# start supervisor on boot
-chmod +x /etc/init.d/supervisord
-update-rc.d supervisord defaults
-service supervisord start
+if [ ! -f /etc/supervisord.conf ]; then
+    echo ""
+    echo "+------------------+"
+    echo "| Setup Supervisor |"
+    echo "+------------------+"
+    echo ""
+    echo_supervisord_conf > /etc/supervisord.conf
+    cat /vagrant/script/supervisord > /etc/init.d/supervisord
+    # add custom tweaks to config file
+    sed -i 's/;chmod=0700/chmod=0764/' /etc/supervisord.conf
+    sed -i 's/;chown=nobody:nogroup/chown=root:vagrant/' /etc/supervisord.conf
+    sed -i 's/;\[include\]/\[include\]/' /etc/supervisord.conf
+    echo "files = /home/vagrant/supervisor/*.conf" >> /etc/supervisord.conf
+    su - vagrant -c 'mkdir ~/supervisor'
+    # start supervisor on boot
+    chmod +x /etc/init.d/supervisord
+    update-rc.d supervisord defaults
+    service supervisord start
+fi
 
 # setup the mysql
-echo ""
-echo "+-------------+"
-echo "| Setup MySQL |"
-echo "+-------------+"
-echo ""
-debconf-set-selections <<< 'mysql-server-5.5 mysql-server/root_password password snapable12345'
-debconf-set-selections <<< 'mysql-server-5.5 mysql-server/root_password_again password snapable12345'
-apt-get -y install mysql-server
+dpkg-query -l mysql-server > /dev/null 2>&1
+INSTALLED=$?
+if [ $INSTALLED != '0' ]; then
+    echo ""
+    echo "+-------------+"
+    echo "| Setup MySQL |"
+    echo "+-------------+"
+    echo ""
+    debconf-set-selections <<< 'mysql-server-5.5 mysql-server/root_password password snapable12345'
+    debconf-set-selections <<< 'mysql-server-5.5 mysql-server/root_password_again password snapable12345'
+    apt-get -y install mysql-server
+    mysql -u root --password=snapable12345 -e 'CREATE DATABASE snapabledb;'
+    mysql -u root --password=snapable12345 -e 'GRANT ALL PRIVILEGES ON snapabledb.* to root@localhost;'
+fi
 
 # setup the snapable api code
 echo ""
@@ -53,13 +61,12 @@ echo "+----------------+"
 echo "| Setup Snapable |"
 echo "+----------------+"
 echo ""
-# setup the mysql db
-mysql -u root --password=snapable12345 -e 'CREATE DATABASE snapabledb;'
-mysql -u root --password=snapable12345 -e 'GRANT ALL PRIVILEGES ON snapabledb.* to root@localhost;'
 # setup the virtualenv
-su - vagrant -c 'mkdir ~/environments'
-su - vagrant -c 'virtualenv -q ~/environments/api'
-su - vagrant -c 'ln -s /vagrant ~/environments/api/snapable'
+if [ ! -d /home/vagrant/environments ]; then
+    su - vagrant -c 'mkdir ~/environments'
+    su - vagrant -c 'virtualenv -q ~/environments/api'
+    su - vagrant -c 'ln -s /vagrant ~/environments/api/snapable'
+fi
 # run the setup instruction commands for the api
 su - vagrant -c '~/environments/api/bin/pip install -v -r /vagrant/requirements.txt'
 su - vagrant -c '~/environments/api/bin/pip install -v -r /vagrant/requirements-dev.txt'
